Simplify hosted voice memo runtime and egress seams
