+++ packages/assistant-engine/src/assistant/codex-runtime.ts
@@ -124,7 +124,6 @@ export async function executeCodexAssistantTurnFromInput(
     activeTurnSessionId: input.activeTurnSessionId,
     conversationHistoryMessages: input.conversationHistoryMessages,
     developerInstructions: input.developerInstructions,
-    dynamicToolRuntime: input.dynamicToolRuntime,
     env: input.env,
     generatedImageUploader: input.generatedImageUploader,
     onCodexThreadHistoryUnsafe: input.onCodexThreadHistoryUnsafe,
@@ -167,7 +166,6 @@ export async function executeCodexAssistantTurnAttemptFromInput(
     activeTurnSessionId: input.activeTurnSessionId,
     conversationHistoryMessages: input.conversationHistoryMessages,
     developerInstructions: input.developerInstructions,
-    dynamicToolRuntime: input.dynamicToolRuntime,
     env: input.env,
     generatedImageUploader: input.generatedImageUploader,
     onCodexThreadHistoryUnsafe: input.onCodexThreadHistoryUnsafe,
