+++ packages/assistant-engine/src/assistant/codex-turn-runner.ts
@@ -436,8 +436,6 @@ async function executeAssistantCodexAttempt(input: {
         executionPlan.onCodexThreadHistoryUnsafe ?? null,
       onFinishWithoutReplyAccepted:
         executionPlan.onFinishWithoutReplyAccepted ?? null,
-      dynamicToolRuntime:
-        executionPlan.input.turnEnvironment?.dynamicToolRuntime ?? null,
       onProviderRequestStarted: (event) => {
         notifyProviderRequestStartedBestEffort({
           event: {
